Graduation Project
This article has been prepared by targeting mathematicians who want to see what can be done by synthesizing selected topics from some of the courses taught in the mathematics departments of universities with computers. While dealing with these issues, the python programming language, which has many advantages, was used. One of the most important benefits of programming is that the computer does the calculations for you. This is necessary when the calculation is long and has many repetitive steps. This article will use our computers to explore mathematical topics that are extremely difficult and time-consuming to solve without software. While exploring these topics, the topics selected from Calculus, Abstract Mathematics, Linear Algebra, Differential Equations, Numerical Analysis, Fourier Analysis, Graph Theory, Fractal Geometry, Coding Theory, Number Theory and Cryptography courses, which are familiar to all mathematicians, respectively, are discussed. It should be noted that knowledge of Python will be needed to understand some code. The paper aims to make learning programming and math more exciting, fun, and rewarding. As a conclusion of this literature survey, demonstrating the application of programming skills to the abstract world of mathematics is highly valuable for the practical solutions of our future problems in the real world.